linter: introduce forbidigo linter to catch fmt.Print* and cleanup
