+++ .golangci.yml
@@ -9,7 +9,22 @@ linters:
     - govet
     - unused
     - nolintlint
+    - forbidigo
 linters-settings:
+  forbidigo:
+    # Forbid the following identifiers (list of regexp).
+    # Default: ["^(fmt\\.Print(|f|ln)|print|println)$"]
+    forbid:
+      # Built-in bootstrapping functions.
+      - ^print(ln)?$
+      # Optional message that gets included in error reports.
+      - p: ^fmt\.Print.*$
+        msg: Do not commit print statements.      
+    # Exclude godoc examples from forbidigo checks.
+    # Default: true
+    exclude-godoc-examples: true
+
+
   exhaustive:
     # Presence of "default" case in switch statements satisfies exhaustiveness,
     # even if all enum members are not listed.
@@ -73,6 +88,22 @@ issues:
         - unused
       path: adapters/repos/db/vector/hnsw/distancer/asm/l2_inline.go
       text: "func `l2[35]`"
+    - linters: # exclude tests
+        - forbidigo
+      path: test/*
+    - linters: # exclude tools
+        - forbidigo
+      path: tools/*
+    - linters: # exclude file
+        - forbidigo
+      path: cmd/libweaviate/libweaviate.go
+    - linters: # exclude file
+        - forbidigo
+      path: adapters/repos/db/vector/hnsw/debug.go
+    - linters: # exclude file
+        - forbidigo
+      path: adapters/repos/db/vector/.*/dump.go
+
 
 run:
   timeout: 5m
