+++ adapters/handlers/rest/clusterapi/indices_replicas.go
@@ -24,6 +24,7 @@ import (
 	"time"
 
 	"github.com/go-openapi/strfmt"
+
 	enterrors "github.com/weaviate/weaviate/entities/errors"
 	"github.com/weaviate/weaviate/entities/storobj"
 	"github.com/weaviate/weaviate/usecases/objects"
@@ -242,7 +243,6 @@ func (i *replicatedIndices) executeCommitPhase() http.Handler {
 func (i *replicatedIndices) increaseReplicationFactor() http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		args := regxIncreaseRepFactor.FindStringSubmatch(r.URL.Path)
-		fmt.Printf("path: %v, args: %+v", r.URL.Path, args)
 		if len(args) != 2 {
 			http.Error(w, "invalid URI", http.StatusBadRequest)
 			return
