+++ adapters/repos/db/helpers/slow_queries.go
@@ -38,16 +38,10 @@ type BaseSlowReporter struct {
 }
 
 func NewSlowQueryReporterFromEnv(logger logrus.FieldLogger) SlowQueryReporter {
-	if logger == nil {
-		fmt.Println("Unexpected nil logger for SlowQueryReporter. Reporter disabled.")
-		return &NoopSlowReporter{}
-	}
-
 	enabled := false
 	if enabledStr, ok := os.LookupEnv(enabledEnvVar); ok {
 		// TODO: Log warning if bool can't be parsed
 		enabled, _ = strconv.ParseBool(enabledStr)
-		fmt.Println("en", enabledStr, enabled)
 	}
 	if !enabled {
 		return &NoopSlowReporter{}
